change all ids from i64 to i32

+++ src/database/audios.rs
@@ -3,21 +3,21 @@ use sqlx::{FromRow, PgPool};
 
 #[derive(FromRow)]
 pub struct DbAudio {
-    pub id: i64,
+    pub id: i32,
     pub transcription: Option<String>,
     pub created_at: DateTime<Utc>,
-    pub user_id: i64,
+    pub user_id: i32,
 }
 
-pub async fn get_audios_by(pool: &PgPool, user_id: i64) -> sqlx::Result<Vec<DbAudio>> {
+pub async fn get_audios_by(pool: &PgPool, user_id: i32) -> sqlx::Result<Vec<DbAudio>> {
     sqlx::query_as("select id, transcription, created_at, user_id from audios where user_id = $1")
         .bind(user_id)
         .fetch_all(pool)
         .await
 }
 
-pub async fn new_audio_by(pool: &PgPool, user_id: i64) -> sqlx::Result<i64> {
-    let id: (i64,) = sqlx::query_as("insert into audios(user_id) values ($1) returning id")
+pub async fn new_audio_by(pool: &PgPool, user_id: i32) -> sqlx::Result<i32> {
+    let id: (i32,) = sqlx::query_as("insert into audios(user_id) values ($1) returning id")
         .bind(user_id)
         .fetch_one(pool)
         .await?;
