+++ src/models/mod.rs
@@ -3,13 +3,13 @@ use serde::Serialize;
 
 #[derive(Serialize)]
 pub struct User {
-    pub id: i64,
+    pub id: i32,
     pub username: String,
 }
 
 #[derive(Serialize)]
 pub struct Audio {
-    pub id: i64,
+    pub id: i32,
     pub transcription: Option<String>,
     pub created_at: DateTime<Utc>,
 }
